# coupon-service/application/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/coupon": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coupon"
                ],
                "summary": "get all coupons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filters for data",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-array_dtos_CouponProfileDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-ApiError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coupon"
                ],
                "summary": "insert/update coupon",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "fa"
                        ],
                        "type": "string",
                        "description": "some description",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "Query Params",
                        "name": "params",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/CouponProfileUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-int"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-ApiError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coupon"
                ],
                "summary": "delete coupons by key",
                "parameters": [
                    {
                        "description": "coupan ids",
                        "name": "ids",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-ApiError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ApiError": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "format": "int32",
                    "allOf": [
                        {
                            "$ref": "#/definitions/core.ErrCode"
                        }
                    ]
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "x-nullable": true,
                    "x-omitempty": true
                }
            }
        },
        "ApiResponse-ApiError": {
            "type": "object",
            "properties": {
                "payload": {
                    "$ref": "#/definitions/ApiError"
                }
            }
        },
        "ApiResponse-array_dtos_CouponProfileDetail": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CouponProfileDetail"
                    }
                }
            }
        },
        "ApiResponse-int": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "integer"
                }
            }
        },
        "CouponProfileUpdate": {
            "type": "object"
        },
        "core.ErrCode": {
            "type": "integer",
            "enum": [
                1001,
                1002
            ],
            "x-enum-varnames": [
                "CouponNotFound",
                "CouponInUse"
            ]
        },
        "dtos.CouponProfileDetail": {
            "type": "object",
            "properties": {
                "couponPluNo": {
                    "$ref": "#/definitions/dtos.NullInt64"
                },
                "couponUpc": {
                    "type": "integer"
                },
                "deptNo": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "modifyDate": {
                    "type": "integer"
                },
                "modifyUser": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "dtos.NullInt64": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}